+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [#2156](https://github.com/influxdb/influxdb/pull/2156): Propagate error when resolving UDP address in Graphite UDP server.
 - [#2163](https://github.com/influxdb/influxdb/pull/2163): Fix up paths for default data and run storage.
 - [#2164](https://github.com/influxdb/influxdb/pull/2164): Append STDOUT/STDERR in initscript.
+- [#2167](https://github.com/influxdb/influxdb/pull/2167): Add broker log recovery.
 
 ## v0.9.0-rc19 [2015-04-01]